Move VerifyImplementation out of the source code i.e AquaHash class.
