+++ unittests/hash_function.cpp
@@ -1,5 +1,6 @@
 #define DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN
 #include "aquahash.h"
+#include "verification.h"
 #include "doctest/doctest.h"
 #include "fmt/format.h"
 #include <string>
@@ -9,20 +10,26 @@
 
 template <typename T> bool test_pod_type(T x, T y) {
     const __m128i seed = _mm_setzero_si128();
-    return _mm_test_all_ones(
-        _mm_cmpeq_epi8(AquaHash::Hash((uint8_t *)(&x), sizeof(T), seed), AquaHash::Hash((uint8_t *)(&y), sizeof(T), seed)));
+    return _mm_test_all_ones(_mm_cmpeq_epi8(AquaHash::Hash((uint8_t *)(&x), sizeof(T), seed),
+                                            AquaHash::Hash((uint8_t *)(&y), sizeof(T), seed)));
 }
 
 bool test_string_type(const std::string &first, const std::string &second) {
     const __m128i seed = _mm_setzero_si128();
-    return _mm_test_all_ones(_mm_cmpeq_epi8(AquaHash::Hash((uint8_t *)(first.data()), first.size(), seed),
-                                            AquaHash::Hash((uint8_t *)(second.data()), second.size(), seed)));
+    return _mm_test_all_ones(
+        _mm_cmpeq_epi8(AquaHash::Hash((uint8_t *)(first.data()), first.size(), seed),
+                       AquaHash::Hash((uint8_t *)(second.data()), second.size(), seed)));
 }
 
 template <typename T> bool test_array_type(const std::vector<T> &first, const std::vector<T> &second) {
     const __m128i seed = _mm_setzero_si128();
-    return _mm_test_all_ones(_mm_cmpeq_epi8(AquaHash::Hash((uint8_t *)(first.data()), first.size() * sizeof(T), seed),
-                                            AquaHash::Hash((uint8_t *)(second.data()), second.size() * sizeof(T), seed)));
+    return _mm_test_all_ones(
+        _mm_cmpeq_epi8(AquaHash::Hash((uint8_t *)(first.data()), first.size() * sizeof(T), seed),
+                       AquaHash::Hash((uint8_t *)(second.data()), second.size() * sizeof(T), seed)));
+}
+
+TEST_CASE("Built-in") {
+    aquahash::VerifyImplementation();
 }
 
 TEST_CASE("POD") {
